requirements.txt
djangorestframework
Django
Markdown
coreapi
gunicorn
PyYAML